+++ src/MathUtil.cpp
@@ -2,21 +2,21 @@
 #include "mathutil.h"
 #include <cmath>
 
-Vector2D mathutil::normaliseVector(Vector2D v, double* magnitude)
+Vector2D mathutil::normaliseVector(Vector2D v, float* magnitude)
 {
-	double m = std::sqrt(std::pow(v.x, 2) + std::pow(v.y, 2));
+	float m = std::sqrt(std::pow(v.x, 2) + std::pow(v.y, 2));
 	Vector2D v_out = { v.x / m, v.y / m };
 	if (magnitude)
 		*magnitude = m;
 	return v_out;
 }
 
-double mathutil::round(double value, int sf)
+float mathutil::round(float value, int sf)
 {
 	return std::floor((value*std::pow(10, sf)) + 0.5) / std::pow(10, sf);
 }
 
-double mathutil::abswrap(double n, double limit)
+float mathutil::abswrap(float n, float limit)
 {
 	if (n < 0)
 		return limit + std::fmod(n, limit);
