+++ src/motionhistory.h
@@ -2,9 +2,9 @@
 
 namespace motionhistory
 {
-	void init(unsigned int max_data_points, double threshold_x, double threshold_y);
-	//void setMotionRecordMaxThresholds(double x, double y);
-	void add(double x, double y);
+	void init(unsigned int max_data_points, float threshold_x, float threshold_y);
+	//void setMotionRecordMaxThresholds(float x, float y);
+	void add(float x, float y);
 	const std::vector<Vector2D>& get();
 	Vector2D getThresholds();
 	unsigned int getCurrentIndex();
